Revert "Add NO_MYMETA=>1 flag to most Makefile.PL's."
This reverts commit 576391c9c2c8d97d32b1e91fdae1de8d32e643c6.

Chris made the point that the NO_MYMETA attribute was introduced
in a version of MakeMaker (6.57_02) that is later than our minimum
required version (6.31). Seems likely that this would cause even
more ugliness for users of older EU::MMs. No time for me to fix
properly before holiday and release, so reverting for now. We
can fix properly after PDL 2.4.10 (this is just cosmetic anyway).