Pass ndimx = 7 to dat_shape
The HDS fortran interface for DAT_SHAPE allocates a temporary array "cdims"
of size DAT__MXDIM (7) but then passes the given ndimx value to datShape.
It in turn this allocates a temporary array of the requested size and then
tries to copy ndimx values from there into cdims.  If ndimx > 7 it will
attempt to write past the end of cdims, causing a segmentation fault.